Fix code-style in Elewant\Reporting
