+++ src/Elewant/Reporting/Application/EventSubscribers/NotifyTwitter.php
@@ -10,22 +10,19 @@ use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
 final class NotifyTwitter implements EventSubscriberInterface
 {
-    /**
-     * @var TwitterOAuth
-     */
+    /** @var TwitterOAuth */
     private $twitterClient;
 
-    /**
-     * @var bool
-     */
+    /** @var bool */
     private $tweetsAreActive;
 
     public function __construct(TwitterOAuth $twitterClient, bool $activateTweets = false)
     {
-        $this->twitterClient   = $twitterClient;
+        $this->twitterClient = $twitterClient;
         $this->tweetsAreActive = $activateTweets;
     }
 
+    /** @return mixed[] */
     public static function getSubscribedEvents(): array
     {
         return [
@@ -49,6 +46,6 @@ final class NotifyTwitter implements EventSubscriberInterface
             $statistics->numberOfNewElePHPants()
         );
 
-        $this->twitterClient->post("statuses/update", ["status" => $status]);
+        $this->twitterClient->post('statuses/update', ['status' => $status]);
     }
 }
