+++ src/Elewant/Reporting/DomainModel/Statistics/NumberOf.php
@@ -4,11 +4,13 @@ declare(strict_types=1);
 
 namespace Elewant\Reporting\DomainModel\Statistics;
 
+use DateTimeInterface;
+
 interface NumberOf
 {
-    public function newHerdsFormedBetween(\DateTimeInterface $from, \DateTimeInterface $to): int;
+    public function newHerdsFormedBetween(DateTimeInterface $from, DateTimeInterface $to): int;
 
     public function herdsEverFormed(): int;
 
-    public function newElePHPantsAdoptedBetween(\DateTimeInterface $from, \DateTimeInterface $to): int;
+    public function newElePHPantsAdoptedBetween(DateTimeInterface $from, DateTimeInterface $to): int;
 }
