+++ src/Elewant/Reporting/Infrastructure/Doctrine/Statistics/NumberOfRepository.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Elewant\Reporting\Infrastructure\Doctrine\Statistics;
 
+use DateTimeInterface;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\NonUniqueResultException;
 use Elewant\Reporting\DomainModel\Statistics\NumberOf;
 
 final class NumberOfRepository implements NumberOf
 {
-    /**
-     * @var EntityManagerInterface
-     */
+    /** @var EntityManagerInterface */
     private $entityManager;
 
     public function __construct(EntityManagerInterface $entityManager)
@@ -20,36 +20,37 @@ final class NumberOfRepository implements NumberOf
     }
 
     /**
-     * @param \DateTimeInterface $from
-     * @param \DateTimeInterface $to
-     *
+     * @param DateTimeInterface $from
+     * @param DateTimeInterface $to
      * @return int
-     * @throws \Doctrine\ORM\NonUniqueResultException
+     * @throws NonUniqueResultException
      */
-    public function newHerdsFormedBetween(\DateTimeInterface $from, \DateTimeInterface $to): int
+    public function newHerdsFormedBetween(DateTimeInterface $from, DateTimeInterface $to): int
     {
-        $dql = <<<EOQ
+        $dql = <<<'EOQ'
 SELECT COUNT(h)
 FROM Herding:Herd h
 WHERE h.formedOn BETWEEN :from AND :to
 EOQ;
 
         $query = $this->entityManager->createQuery($dql);
-        $query->setParameters([
-            'from' => $from->format('Y-m-d 00:00:00'),
-            'to'   => $to->format('Y-m-d 23:59:59'),
-        ]);
+        $query->setParameters(
+            [
+                'from' => $from->format('Y-m-d 00:00:00'),
+                'to' => $to->format('Y-m-d 23:59:59'),
+            ]
+        );
 
         return (int) $query->getSingleScalarResult();
     }
 
     /**
      * @return int
-     * @throws \Doctrine\ORM\NonUniqueResultException
+     * @throws NonUniqueResultException
      */
     public function herdsEverFormed(): int
     {
-        $dql = <<<EOQ
+        $dql = <<<'EOQ'
 SELECT COUNT(h)
 FROM Herding:Herd h
 EOQ;
@@ -60,25 +61,26 @@ EOQ;
     }
 
     /**
-     * @param \DateTimeInterface $from
-     * @param \DateTimeInterface $to
-     *
+     * @param DateTimeInterface $from
+     * @param DateTimeInterface $to
      * @return int
-     * @throws \Doctrine\ORM\NonUniqueResultException
+     * @throws NonUniqueResultException
      */
-    public function newElePHPantsAdoptedBetween(\DateTimeInterface $from, \DateTimeInterface $to): int
+    public function newElePHPantsAdoptedBetween(DateTimeInterface $from, DateTimeInterface $to): int
     {
-        $dql = <<<EOQ
+        $dql = <<<'EOQ'
 SELECT COUNT(e)
 FROM Herding:ElePHPant e
 WHERE e.adoptedOn BETWEEN :from AND :to
 EOQ;
 
         $query = $this->entityManager->createQuery($dql);
-        $query->setParameters([
-            'from' => $from->format('Y-m-d 00:00:00'),
-            'to'   => $to->format('Y-m-d 23:59:59'),
-        ]);
+        $query->setParameters(
+            [
+                'from' => $from->format('Y-m-d 00:00:00'),
+                'to' => $to->format('Y-m-d 23:59:59'),
+            ]
+        );
 
         return (int) $query->getSingleScalarResult();
     }
